renaming shell ->  interactive
