+++ src/command_modules/azure-cli-interactive/MANIFEST.in
