+++ src/command_modules/azure-cli-interactive/azclishell/telemetry.py
