+++ src/command_modules/azure-cli-interactive/azure/cli/command_modules/interactive/_params.py
@@ -3,9 +3,11 @@
 # Licensed under the MIT License. See License.txt in the project root for license information.
 # --------------------------------------------------------------------------------------------
 
-from azure.cli.core.help_files import helps
+# pylint: disable=line-too-long
 
-helps['shell'] = """
-            type: command
-            short-summary: Start the interactive shell for Microsoft Azure CLI (preview)
-            """
+from azure.cli.core.commands import register_cli_argument
+
+from azclishell.color_styles import get_options as style_options
+
+register_cli_argument('launch-interactive', 'style', options_list=('--style', '-s'),
+                      help='The colors of the shell.', choices=style_options())
