+++ src/command_modules/azure-cli-interactive/docs/faster-shell-cast.gif
