+++ src/command_modules/azure-cli-interactive/docs/shell.gif
