+++ src/command_modules/azure-cli-interactive/setup.cfg
