+++ src/command_modules/azure-cli-interactive/tests/test_completion.py
