+++ src/command_modules/azure-cli-interactive/tests/test_gather.py
