+++ src/command_modules/azure-cli-interactive/tests/test_quote_parse.py
